% Generated by roxygen2 (4.0.0): do not edit by hand
\name{angle}
\alias{angle}
\title{Compute the angle between two lines}
\usage{
angle(A, B)
}
\arguments{
  \item{A}{First line}

  \item{B}{Second line}
}
\value{
Angle in radians (scalar)
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

