% Generated by roxygen2 (4.0.0): do not edit by hand
\name{center}
\alias{center}
\title{Center a line by translating so the start point is the origin}
\usage{
center(line)
}
\value{
Object of Line class
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{angle}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

