% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Circle}
\alias{Circle}
\title{Circle class}
\usage{
Circle(center, radius)
}
\arguments{
  \item{center}{Center of the circle (Point)}

  \item{radius}{Radius of the circle (scalar)}
}
\value{
Object of Circle class
}
\description{
Internal class for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{IntersectLC}};
\code{\link{Line}}; \code{\link{Point}};
\code{\link{RefractLC}}; \code{\link{RefractLL}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

