% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Constants}
\alias{Constants}
\title{Constants for IOL calculations and biometry}
\format{\preformatted{List of 3
 $ IOL     :List of 8
  ..$ A_to_pACD_a0: num -63.9
  ..$ A_to_pACD_a1: num 0.584
  ..$ A_to_S_a0   : num -65.6
  ..$ A_to_S_a1   : num 0.566
  ..$ pACD_to_S_a0: num -3.6
  ..$ pACD_to_S_a1: num 0.97
  ..$ A_to_ACD_a0 : num -68.7
  ..$ A_to_ACD_a1 : num 0.625
 $ Average :List of 10
  ..$ L                : num 23.5
  ..$ L.sd             : num 1.25
  ..$ K                : num 43.8
  ..$ K.sd             : num 1.6
  ..$ ACD              : num 3.1
  ..$ ACD.sd           : num 0.3
  ..$ lens.thickness   : num 4.7
  ..$ lens.thickness.sd: num 0.41
  ..$ age              : num 72
  ..$ age.sd           : num 12
 $ Biometry:List of 14
  ..$ us_to_pci              : num 0.15
  ..$ us_1532_to_us          : num 0.32
  ..$ us_applination_to_us   : num 0.2
  ..$ immersion_us_to_pci_a1 : num 0.995
  ..$ immersion_us_to_pci_a0 : num 0.0779
  ..$ us_immersion_error     : num 0.12
  ..$ us_applination_error   : num 0.13
  ..$ us_immersion_1532      : num 0.13
  ..$ pci_iolmaster_error    : num 0.02
  ..$ corneal_index          : num 1.34
  ..$ corneal_index_2        : num 1.33
  ..$ aqueous_index          : num 1.34
  ..$ retinal_thickness      : num 0.2
  ..$ corneal_principle_plane: num 0.05
}}
\usage{
Constants
}
\description{
Constants for IOL calculations and biometry
}
\keyword{datasets}

