% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ELP}
\alias{ELP}
\title{Calculate Effective Lens Position (ELP) from Biometry and IOL Data}
\usage{
ELP(L, K, R = NA, cornea_n = NA, A = NA, pACD = NA, S = NA,
  which = "modern")
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{K}{average corneal curvature of the eye in diopters
  (D)}

  \item{R}{average corneal radius in millimeters (mm)}

  \item{cornea_n}{effective corneal index of refraction}

  \item{A}{IOL A constant}

  \item{pACD}{IOL pACD constant}

  \item{S}{IOL surgeon factor constant}

  \item{which}{string vector specifying which ELP formulas
  to use}
}
\value{
Named numeric vector of effective lens position (in mm) for
each ELP formula requested.
}
\description{
Using specified ocular biometry data including a measure of
the eye's optical length and corneal curvature, this
function calculates one or more estimates of an intraocular
lens's effective lens position (ELP). The ELP is the
position in millimeters of the IOL's principle plane from
the cornea's principle plane. The ELP used to be referred
to as the anterior chamber depth (ACD); however, ELP is a
more accurate description.
}
\note{
If the some of the IOL constants A, pACD, or S are not
provided, it may be derived from those given. A warning is
generally produced when this conversion is performed.

The returned numeric vector is augmented with a
\code{'parameters'} list attribute describing which
biometry and IOL parameters were used to calculate each
value. For example, if the Hoffer Q ELP is calculated given
an axial length, corneal power (via the \code{K}
parameter), and an A-constant (via the \code{A} parameter),
then the \code{'parameters'} list attribute will have a
list named \code{Hoffer.Q} with three values: \code{L},
\code{K}, and \code{A}.
}
\examples{
# Get the effective lens position of a normal eye with for the
# Alcon SA60AT lens using the Hoffer Q formula. This will compute
# the required pACD IOL constant from the provided A constant.
(elp <- ELP(L = 24, K = 44, A = 118.4, which = 'Hoffer.Q'))

# Check which parameters were used to calculate the ELP
attr(elp, 'parameters')$Hoffer.Q

# Get the ELP of a normal eye with modern formulas (Hoffer Q, SRK/T,
# and Holladay 1). Five warnings will be output for the conversion of
# the A constant to ACD for the SRK/T and Hoffer Q formulas, using a
# standard corneal index of refraction (since one wasn't provided) to
# convert corneal power to radius of curvature, and for approximating
# the Holladay 1 surgeon factor from the provided A constant.
(elp <- ELP(L = 24, K = 44, A = 118.4, which = 'modern'))

# Get the ELP of a normal eye with the Holladay 1 formula. Although
# both the IOL A constant and surgeon factor are provided, since the
# formula requires the surgeon factor, the A constant will be ignored
(elp <- ELP(L = 24, K = 44, A = 118.4, S = 1.45, which = 'Holladay.1'))
attr(elp, 'parameters')$Holladay.1
}
\author{
Eric N. Brown \email{eric.n.brown@gmail.com}
}
\seealso{
\code{\link{Power}}

Other ELP: \code{\link{Haigis.ELP}};
\code{\link{Hoffer.ELP}}; \code{\link{Hoffer.Q.ELP}};
\code{\link{Holladay.1.ELP}}; \code{\link{Holladay.ELP}};
\code{\link{SRK.T.ELP}}
}

