% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Fyodorov.Power}
\alias{Fyodorov.Power}
\title{Fyodorov Formula for IOL Power}
\usage{
Fyodorov.Power(L, K, ELP)
}
\arguments{
  \item{L}{axial length in millimeters (mm)}

  \item{K}{corneal curvature in diopters (D)}

  \item{ELP}{effective lens position in millimeters (mm)}
}
\value{
IOL power for emmetropia
}
\description{
Calculate IOL power for emmetropia given axial length,
corneal power, and effective lens position.
}
\references{
\url{https://encrypted.google.com/books?id=NhWJsGFK6qgC&pg=PA8}
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{SRK.T.Power}}; \code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

