% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Haigis.ELP}
\alias{Haigis.ELP}
\title{Haigis Formula for Effective Lens Position}
\usage{
Haigis.ELP(pACD)
}
\arguments{
  \item{pACD}{personalized ACD (mm) lens constant}
}
\value{
ELP in mm
}
\description{
Calculate IOL effective lens position for emmetropia given
personalized lens-constant anterior chamber depth (pACD).
}
\details{
Note: Currently this is just the provided pACD constant.
}
\seealso{
\code{\link{ELP}}

Other ELP: \code{\link{ELP}}; \code{\link{Hoffer.ELP}};
\code{\link{Hoffer.Q.ELP}}; \code{\link{Holladay.1.ELP}};
\code{\link{Holladay.ELP}}; \code{\link{SRK.T.ELP}}
}

