% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Hoffer.Power}
\alias{Hoffer.Power}
\title{Hoffer Formula for Emmetropic IOL Power}
\usage{
Hoffer.Power(L, K, ELP)
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{K}{average corneal curvature of the eye in diopters
  (D)}

  \item{ELP}{IOL effective lens position (mm)}
}
\value{
Power of emmetropic IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and effective lens position.
}
\references{
\url{https://encrypted.google.com/books?id=NhWJsGFK6qgC&pg=PA8}
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{SRK.T.Power}}; \code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

