% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Holladay.1.ELP}
\alias{Holladay.1.ELP}
\title{Holladay 1 Formula for Effective Lens Position}
\usage{
Holladay.1.ELP(L, R, K, cornea_n, S, A, pACD)
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{R}{corneal radius of curvature (mm)}

  \item{K}{corneal power (D) used to determine corneal
  radius of curvature}

  \item{cornea_n}{corneal index of refraction used to
  determine corneal radius of curvature from K}

  \item{S}{IOL surgeon-factor constant (mm)}

  \item{A}{IOL A constant (D) used to determine
  surgeon-factor}

  \item{pACD}{IOL personalized ACD constant (mm) used to
  determine surgeon-factor}
}
\value{
ELP in mm
}
\description{
Calculate IOL effective lens position for emmetropia given
axial length, corneal radius of curvature, and lens
constant.
}
\details{
Note: If the Holladay Surgeon-factor constant for the lens
is not provided, it is calculated from the lens A-constant
or pACD-constant. Similarily, if the corneal radius of
curvature is not provided, it is calculated from the
corneal curvature (in diopters) and corneal index of
refraction.
}
\references{
\url{https://encrypted.google.com/books?id=NhWJsGFK6qgC&pg=PA8}
}
\seealso{
\code{\link{ELP}}

Other ELP: \code{\link{ELP}}; \code{\link{Haigis.ELP}};
\code{\link{Hoffer.ELP}}; \code{\link{Hoffer.Q.ELP}};
\code{\link{Holladay.ELP}}; \code{\link{SRK.T.ELP}}
}

