% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Holladay.1.Power}
\alias{Holladay.1.Power}
\title{Holladay-1 Formula for IOL Power}
\usage{
Holladay.1.Power(L, R, ELP, K, cornea_n = 1.3375, aqueous_n = 1.336,
  RT = 0.2, Rx = 0, V = 13)
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{R}{corneal radius of curvature (mm)}

  \item{ELP}{IOL effective lens position (mm)}

  \item{K}{corneal power (D) used to determine corneal
  radius of curvature}

  \item{cornea_n}{corneal index of refraction used to
  determine corneal radius of curvature from K, defaults to
  1.3375}

  \item{aqueous_n}{aqueous index of refraction, defaults to
  1.3375}

  \item{RT}{retinal thickness (mm), defaults to 0.2 mm}

  \item{Rx}{desired resulting refractive error (D),
  defaults to 0 D}

  \item{V}{vertex distance of Rx (mm), defaults to 13 mm}
}
\value{
Power of IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and effective lens position.
}
\details{
Note: if the corneal radius, \code{R}, is not provided, it
is determined from the corneal power, \code{K}, and the
corneal index of refraction, \code{cornea_n}.
}
\references{
\url{https://encrypted.google.com/books?id=NhWJsGFK6qgC&pg=PA8}
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{SRK.T.Power}}; \code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

