% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Holladay.ELP}
\alias{Holladay.ELP}
\title{Holladay Formula for Effective Lens Position}
\usage{
Holladay.ELP(A)
}
\arguments{
  \item{A}{IOL A-constant (D)}
}
\value{
ELP in mm
}
\description{
Calculate IOL effective lens position for emmetropia given
lens A-constant.
}
\references{
\url{http://books.google.com/books?id=SHjgbQ9auqYC&pg=PT93}
}
\seealso{
\code{\link{ELP}}

Other ELP: \code{\link{ELP}}; \code{\link{Haigis.ELP}};
\code{\link{Hoffer.ELP}}; \code{\link{Hoffer.Q.ELP}};
\code{\link{Holladay.1.ELP}}; \code{\link{SRK.T.ELP}}
}

