% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Holladay.Power}
\alias{Holladay.Power}
\title{Holladay Formula for IOL Power}
\usage{
Holladay.Power(L, K, ELP, Rx = 0, V = 13)
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{K}{corneal power (D)}

  \item{ELP}{IOL effective lens position (mm)}

  \item{Rx}{desired resulting refractive error (D),
  defaults to 0 D}

  \item{V}{vertex distance of Rx (mm), defaults to 13 mm}
}
\value{
Power of IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and effective lens position.
}
\references{
\url{http://books.google.com/books?id=SHjgbQ9auqYC&pg=PT93}
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{SRK.T.Power}}; \code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

