% Generated by roxygen2 (4.0.0): do not edit by hand
\name{IntersectLC}
\alias{IntersectLC}
\title{Intersect and Line and Circle}
\usage{
IntersectLC(line, circle)
}
\arguments{
  \item{line}{Line to intersect (treated as an infinite
  line)}

  \item{circle}{Circle to intersect}
}
\value{
A list of 0, 1, or 2 Points of intersection between the
line and circle
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLV}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{RefractThinLens}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

