% Generated by roxygen2 (4.0.0): do not edit by hand
\name{IntersectLV}
\alias{IntersectLV}
\title{Intersect and Line and a vertical line}
\usage{
IntersectLV(line, position)
}
\arguments{
  \item{line}{Line to intersect (treated as an infinite
  line)}

  \item{position}{Position of vertical line (x-coordinate)
  to intersect}
}
\value{
Point of intersection between the line and vertical line
(or NA)
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{RefractThinLens}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

