% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{IOL-package}
\alias{IOL}
\alias{IOL-package}
\title{Calculates ophthalmic intraocular lens powers based on biometry and
  lens parameters.}
\description{
A package to calculate ophthalmologic intraocular lens
(IOL) powers.
}
\author{
Eric N. Brown \email{eric.n.brown@gmail.com}
}
\seealso{
\code{\link{ELP}}, \code{\link{Power}}
}
\keyword{package}

