% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Line}
\alias{Line}
\title{Line class}
\usage{
Line(a, b)
}
\arguments{
  \item{A}{Starting point (of Point class)}

  \item{B}{Ending point (of Point class)}
}
\value{
Object of Line class
}
\description{
Internal class for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{RefractThinLens}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

