% Generated by roxygen2 (4.0.0): do not edit by hand
\name{orthogonal}
\alias{orthogonal}
\title{Return an orthogonal line}
\usage{
orthogonal(line)
}
\value{
Object of Line class
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Line}}; \code{\link{Point}};
\code{\link{RefractLC}}; \code{\link{RefractLL}};
\code{\link{RefractThinLens}}; \code{\link{angle}};
\code{\link{center}}; \code{\link{length.Line}};
\code{\link{normalize}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

