% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.Circle}
\alias{plot.Circle}
\title{Plot a Circle class on an existing plot}
\usage{
\method{plot}{Circle}(c, points = 100, around = NULL, ...)
}
\arguments{
  \item{points}{Number of points with which to approximate
  a circle}

  \item{around}{Optional Point or Line to limit the extent
  of the drawn circle}

  \item{...}{Other options to pass to the base graphics
  lines function}
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Line}}; \code{\link{Point}};
\code{\link{RefractLC}}; \code{\link{RefractLL}};
\code{\link{RefractThinLens}}; \code{\link{angle}};
\code{\link{center}}; \code{\link{length.Line}};
\code{\link{normalize}}; \code{\link{orthogonal}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

