% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.Line}
\alias{plot.Line}
\title{Plot a Line class on an existing plot}
\usage{
\method{plot}{Line}(ln, ...)
}
\arguments{
  \item{...}{Other options to pass to the base graphics
  segments function}
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Line}}; \code{\link{Point}};
\code{\link{RefractLC}}; \code{\link{RefractLL}};
\code{\link{RefractThinLens}}; \code{\link{angle}};
\code{\link{center}}; \code{\link{length.Line}};
\code{\link{normalize}}; \code{\link{orthogonal}};
\code{\link{plot.Circle}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

