% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Point}
\alias{Point}
\title{Point class}
\usage{
Point(x, y)
}
\arguments{
  \item{x}{x coordinate}

  \item{y}{y coordinate}
}
\value{
Object of Point class
}
\description{
Internal class for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Line}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{RefractThinLens}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

