% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Power}
\alias{Power}
\title{Calculate IOL Power from Biometry Data and ELP}
\usage{
Power(L, K, A, ELP, Rx = 0, V = 13, which = "all")
}
\arguments{
  \item{L}{length of the eye in millimeters (mm)}

  \item{K}{average corneal curvature of the eye in diopters
  (D)}

  \item{A}{IOL A constant}

  \item{ELP}{effective lens position in millimeters (mm)}

  \item{Rx}{desired refractive outcome in diopters (D),
  defaults to emmetropia}

  \item{V}{vertex of refractive outcome in millimeters
  (mm), defaults to 13 mm}

  \item{which}{string vector specifying which IOL power
  formulas to use}
}
\value{
Named numeric vector of optimal IOL powers (in diopters, D)
for each formula requested
}
\description{
Using specified ocular biometry data including a measure of
the eye's optical length and corneal curvature, this
function calculates one or more estimates of an intraocular
lens's effective lens position (ELP). The ELP is the
position in millimeters of the IOL's principle plane from
the cornea's principle plane. The ELP used to be referred
to as the anterior chamber depth (ACD); however, ELP is a
more accurate description.
}
\author{
Eric N. Brown \email{eric.n.brown@gmail.com}
}
\seealso{
\code{\link{ELP}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{SRK.II.Power}};
\code{\link{SRK.Power}}; \code{\link{SRK.T.Power}};
\code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

