% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RefractLC}
\alias{RefractLC}
\title{Refract a line into a circle}
\usage{
RefractLC(line, circle, n1 = 1, n2 = 1, first = TRUE, inside = !first,
  plot = TRUE, ...)
}
\arguments{
  \item{line}{Infinite line to refract}

  \item{normal}{Normal of the refracting surface (the
  surface is located at the normal's starting point)}

  \item{n1}{Index of refraction for the incident ray}

  \item{n2}{Index of refraction for the refracted ray}

  \item{first}{Whether to refract at the first intersection
  of the line and circle or the second.}

  \item{inside}{Whether the incident line is leaving the
  circle (starting inside) and the refracting ray is
  outside of the circle}

  \item{plot}{Whether to plot the incident ray, surface
  normal, and refracted ray}
}
\value{
The refracted ray with the starting point being the
normal's starting point (class Line)
}
\description{
Internal function for ray-tracing 2D optics.
}
\details{
Refracts a line into a circle by first computing the
intersection of the line and the circle and then computing
the surface normal of the circle at that intersection
point. The parameters first and inside choose whether the
first (left/top) or second (right/bottom) intersection of
the line with the circle is the chosen refraction surface.
The inside parameter determines whether the refraction is
taken as a ray intersecting in-to or out-of the circle.
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{IntersectLV}};
\code{\link{Line}}; \code{\link{Point}};
\code{\link{RefractLL}}; \code{\link{RefractThinLens}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

