% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RefractLL}
\alias{RefractLL}
\title{Refract a line given a normal direction and two index of refractions}
\usage{
RefractLL(line, normal, n1 = 1, n2 = 1, plot = TRUE, ...)
}
\arguments{
  \item{line}{Infinite line to refract}

  \item{normal}{Normal of the refracting surface (the
  surface is located at the normal's starting point)}

  \item{n1}{Index of refraction for the incident ray}

  \item{n2}{Index of refraction for the refracted ray}

  \item{plot}{Whether to plot the incident ray, surface
  normal, and refracted ray}
}
\value{
The refracted ray with the starting point being the
normal's starting point (class Line)
}
\description{
Internal function for ray-tracing 2D optics.
}
\details{
The line argument provides the direction for the incident
ray which is assumed to intersect with the normal at the
normal's starting point. This is not checked and only the
direction of the incident line matters (not its location or
scale). The resulting refracted ray also starts at the
normal's starting point.
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{angle}}; \code{\link{center}};
\code{\link{length.Line}}; \code{\link{normalize}};
\code{\link{orthogonal}}; \code{\link{plot.Circle}};
\code{\link{plot.Line}}; \code{\link{plot.Point}};
\code{\link{rotate}}; \code{\link{scale}}
}
\keyword{internal}

