% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rotate}
\alias{rotate}
\title{Rotate a line about the starting point}
\usage{
rotate(line, angle)
}
\arguments{
  \item{angle}{Angle to rotate (in radians)}
}
\value{
Object of Line class
}
\description{
Internal function for ray-tracing 2D optics
}
\seealso{
Other optics: \code{\link{Circle}};
\code{\link{IntersectLC}}; \code{\link{Line}};
\code{\link{Point}}; \code{\link{RefractLC}};
\code{\link{RefractLL}}; \code{\link{angle}};
\code{\link{center}}; \code{\link{length.Line}};
\code{\link{normalize}}; \code{\link{orthogonal}};
\code{\link{plot.Circle}}; \code{\link{plot.Line}};
\code{\link{plot.Point}}; \code{\link{scale}}
}
\keyword{internal}

