% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Shammas.Power}
\alias{Shammas.Power}
\title{Shammas Formula for IOL Power}
\usage{
Shammas.Power(L, K, ELP)
}
\arguments{
  \item{L}{axial length of the eye in millimeters (mm) from
  immersion ultrasound}

  \item{K}{corneal power (D)}

  \item{ELP}{IOL effective lens position (mm)}
}
\value{
Power of IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and effective lens position.
}
\details{
Note: The axial length is assumed to be from immersion
ultrasound. For contact ultrasound, add 0.24 mm.
}
\references{
\url{https://encrypted.google.com/books?id=NhWJsGFK6qgC&pg=PA8}
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{SRK.T.Power}};
\code{\link{van.der.Heijde.Power}}
}

