% Generated by roxygen2 (4.0.0): do not edit by hand
\name{SRK.II.Power}
\alias{SRK.II.Power}
\title{SRK-II Formula for IOL Power}
\usage{
SRK.II.Power(L, K, A, ELP)
}
\arguments{
  \item{L}{axial length of the eye in millimeters (mm)}

  \item{K}{corneal power (D)}

  \item{A}{IOL A-constant (D)}

  \item{ELP}{IOL effective lens position (mm) used to
  calculate equivalent A-constant}
}
\value{
Power of IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and IOL A-constant.
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Haigis.Power}};
\code{\link{Hoffer.Power}}; \code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.Power}}; \code{\link{SRK.T.Power}};
\code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

