% Generated by roxygen2 (4.0.0): do not edit by hand
\name{SRK.T.Power}
\alias{SRK.T.Power}
\title{SRK/T Formula for IOL Power}
\usage{
SRK.T.Power(L, K, ELP)
}
\arguments{
  \item{L}{axial length of the eye in millimeters (mm)}

  \item{K}{corneal power (D)}

  \item{ELP}{IOL effective lens position (mm)}
}
\value{
Power of IOL (D)
}
\description{
Calculate IOL power for emmetropia given axial length,
corenal curvature, and effective lens position.
}
\seealso{
\code{\link{Power}}

Other Power: \code{\link{Binkhorst.Power}};
\code{\link{Binkhorst.adjusted.Power}};
\code{\link{Colenbrander.Power}};
\code{\link{Fyodorov.Power}}; \code{\link{Hoffer.Power}};
\code{\link{Hoffer.Q.Power}};
\code{\link{Holladay.1.Power}};
\code{\link{Holladay.Power}}; \code{\link{Power}};
\code{\link{SRK.II.Power}}; \code{\link{SRK.Power}};
\code{\link{Shammas.Power}};
\code{\link{van.der.Heijde.Power}}
}

